requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
pygad>=3.0.0
pyyaml>=5.4.0
pytest>=6.2.0
pytest-cov>=2.12.0
plotly>=5.0.0